add admin dashboard

+++ app/helpers/application_helper.rb
@@ -4,8 +4,12 @@ module ApplicationHelper
       (link_to 'Login', new_user_session_path, class: 'main-nav__link') +
       (link_to 'Register', new_user_registration_path, class: 'main-nav__link')
     else
-      (link_to 'Admin', admin_path,class: 'main-nav__link' if logged_in?(:site_admin)) +
-      (link_to 'Logout', destroy_user_session_path, {method: :DELETE, class: 'main-nav__link'})
+      if logged_in?(:site_admin)
+        return (link_to 'Admin', admin_root_path, class: 'main-nav__link') +
+        (link_to 'Logout', destroy_user_session_path, {method: :DELETE, class: 'main-nav__link'})
+      end
+
+      link_to 'Logout', destroy_user_session_path, {method: :DELETE, class: 'main-nav__link'}
     end
   end
